// src/api/models/CreateJourneyDefinitionRequest.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Journey Orchestrator API
 * Platform service responsible for executing configurable journeys using a state machine model. The service is business-agnostic and manages journey definitions and journey instances driven by events. 
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { TransitionRequest } from './TransitionRequest';
import {
    TransitionRequestFromJSON,
    TransitionRequestFromJSONTyped,
    TransitionRequestToJSON,
    TransitionRequestToJSONTyped,
} from './TransitionRequest';
import type { State } from './State';
import {
    StateFromJSON,
    StateFromJSONTyped,
    StateToJSON,
    StateToJSONTyped,
} from './State';

/**
 * 
 * @export
 * @interface CreateJourneyDefinitionRequest
 */
export interface CreateJourneyDefinitionRequest {
    /**
     * 
     * @type {string}
     * @memberof CreateJourneyDefinitionRequest
     */
    journeyCode: string;
    /**
     * 
     * @type {string}
     * @memberof CreateJourneyDefinitionRequest
     */
    name: string;
    /**
     * 
     * @type {number}
     * @memberof CreateJourneyDefinitionRequest
     */
    version: number;
    /**
     * 
     * @type {Array<State>}
     * @memberof CreateJourneyDefinitionRequest
     */
    states: Array<State>;
    /**
     * 
     * @type {Array<TransitionRequest>}
     * @memberof CreateJourneyDefinitionRequest
     */
    transitions: Array<TransitionRequest>;
}

/**
 * Check if a given object implements the CreateJourneyDefinitionRequest interface.
 */
export function instanceOfCreateJourneyDefinitionRequest(value: object): value is CreateJourneyDefinitionRequest {
    if (!('journeyCode' in value) || value['journeyCode'] === undefined) return false;
    if (!('name' in value) || value['name'] === undefined) return false;
    if (!('version' in value) || value['version'] === undefined) return false;
    if (!('states' in value) || value['states'] === undefined) return false;
    if (!('transitions' in value) || value['transitions'] === undefined) return false;
    return true;
}

export function CreateJourneyDefinitionRequestFromJSON(json: any): CreateJourneyDefinitionRequest {
    return CreateJourneyDefinitionRequestFromJSONTyped(json, false);
}

export function CreateJourneyDefinitionRequestFromJSONTyped(json: any, ignoreDiscriminator: boolean): CreateJourneyDefinitionRequest {
    if (json == null) {
        return json;
    }
    return {
        
        'journeyCode': json['journeyCode'],
        'name': json['name'],
        'version': json['version'],
        'states': ((json['states'] as Array<any>).map(StateFromJSON)),
        'transitions': ((json['transitions'] as Array<any>).map(TransitionRequestFromJSON)),
    };
}

export function CreateJourneyDefinitionRequestToJSON(json: any): CreateJourneyDefinitionRequest {
    return CreateJourneyDefinitionRequestToJSONTyped(json, false);
}

export function CreateJourneyDefinitionRequestToJSONTyped(value?: CreateJourneyDefinitionRequest | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'journeyCode': value['journeyCode'],
        'name': value['name'],
        'version': value['version'],
        'states': ((value['states'] as Array<any>).map(StateToJSON)),
        'transitions': ((value['transitions'] as Array<any>).map(TransitionRequestToJSON)),
    };
}