// src/api/models/State.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Journey Orchestrator API
 * Platform service responsible for executing configurable journeys using a state machine model. The service is business-agnostic and manages journey definitions and journey instances driven by events. 
 *
 * The version of the OpenAPI document: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { StateType } from './StateType';
import {
    StateTypeFromJSON,
    StateTypeFromJSONTyped,
    StateTypeToJSON,
    StateTypeToJSONTyped,
} from './StateType';

/**
 * 
 * @export
 * @interface State
 */
export interface State {
    /**
     * 
     * @type {string}
     * @memberof State
     */
    name: string;
    /**
     * 
     * @type {StateType}
     * @memberof State
     */
    type: StateType;
}



/**
 * Check if a given object implements the State interface.
 */
export function instanceOfState(value: object): value is State {
    if (!('name' in value) || value['name'] === undefined) return false;
    if (!('type' in value) || value['type'] === undefined) return false;
    return true;
}

export function StateFromJSON(json: any): State {
    return StateFromJSONTyped(json, false);
}

export function StateFromJSONTyped(json: any, ignoreDiscriminator: boolean): State {
    if (json == null) {
        return json;
    }
    return {
        
        'name': json['name'],
        'type': StateTypeFromJSON(json['type']),
    };
}

export function StateToJSON(json: any): State {
    return StateToJSONTyped(json, false);
}

export function StateToJSONTyped(value?: State | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'name': value['name'],
        'type': StateTypeToJSON(value['type']),
    };
}